+++ src/com/android/phone/NotificationMgr.java
@@ -482,7 +482,7 @@ public class NotificationMgr {
     private void showNetworkSelection(String operator, Phone phone) {
         if (DBG) log("showNetworkSelection(" + operator + ")...");
 
-        long subId = phone.getSubId();
+        int subId = phone.getSubId();
         Notification.Builder builder = new Notification.Builder(mContext)
                 .setSmallIcon(android.R.drawable.stat_sys_warning)
                 .setContentTitle(mContext.getString(R.string.notification_network_selection_title))
