+++ src/com/android/services/telephony/TelephonyConnectionService.java
@@ -584,7 +584,7 @@ public class TelephonyConnectionService extends ConnectionService {
 
         if (phoneId == -1) {
             for (int phId = 0; phId < phoneCount; phId++) {
-                long[] subId = scontrol.getSubId(phId);
+                int[] subId = scontrol.getSubId(phId);
                 if ((tm.getSimState(phId) == TelephonyManager.SIM_STATE_READY) &&
                         (scontrol.getSubState(subId[0]) == SubscriptionManager.ACTIVE)) {
                     phoneId = phId;
